Performance improvement: On 25k rows this makes it 3 seconds faster to load.
Before the visualization would run all queued filters and compute
all processed metrics for all rows. Now we do it only for the
once we display which makes it much faster. Also we no longer have
duplicated code in Visualization::applyFilters with
DataTablePostProcesser::process and we make sure the reports in the
UI are shown the same way and order as in API requests as before
filters were executed in a very random order which let to different
results.